Added Edge Handling and Multithread Processing
Added:
- edge handling (mirror technique, outside pixels are taken from inside)
- processing on four threads
- changed double to float for processing power
- all line should now be under 120 characters

+++ src/ro/damianteodorbeles/imageprocessing/IKernel.java
@@ -2,7 +2,7 @@ package ro.damianteodorbeles.imageprocessing;
 
 public interface IKernel {
 	
-	public double [] getKernel();
+	public float [] getKernel();
 	public int getWidth();
 	public int getHeight();
 	public int [] getOrigin();
