+++ src/ro/damianteodorbeles/imageprocessing/KernelFactory.java
@@ -1,18 +1,14 @@
 package ro.damianteodorbeles.imageprocessing;
 
-
 public class KernelFactory {
-	
+
 	public IKernel BOX_BLUR() {
 		return new IKernel() {
 
 			@Override
-			public double[] getKernel() {
-				return new double[] {
-					1.0/9, 1.0/9, 1.0/9,
-					1.0/9, 1.0/9, 1.0/9,
-					1.0/9, 1.0/9, 1.0/9
-				};
+			public float[] getKernel() {
+				return new float[] { (float) (1) / 9, (float) (1) / 9, (float) (1) / 9, (float) (1) / 9,
+						(float) (1) / 9, (float) (1) / 9, (float) (1) / 9, (float) (1) / 9, (float) (1) / 9 };
 			}
 
 			@Override
@@ -31,17 +27,14 @@ public class KernelFactory {
 			}
 		};
 	}
-	
+
 	public IKernel GAUSSIAN_BLUR() {
 		return new IKernel() {
 
 			@Override
-			public double[] getKernel() {
-				return new double[] {
-						1.0/16, 2.0/16, 1.0/16,
-						2.0/16, 4.0/16, 2.0/16,
-						1.0/16, 2.0/16, 1.0/16
-				};
+			public float[] getKernel() {
+				return new float[] { (float) (1) / 16, (float) (2) / 16, (float) (1) / 16, (float) (2) / 16,
+						(float) (4) / 16, (float) (2) / 16, (float) (1) / 16, (float) (2) / 16, (float) (1) / 16 };
 			}
 
 			@Override
@@ -58,20 +51,16 @@ public class KernelFactory {
 			public int[] getOrigin() {
 				return new int[] { 1, 1 };
 			}
-			
+
 		};
 	}
-	
+
 	public IKernel EDGE_DETECTION() {
 		return new IKernel() {
 
 			@Override
-			public double[] getKernel() {
-				return new double[] {
-					-1, -1, -1,
-					-1,  8, -1,
-					-1, -1, -1
-				};
+			public float[] getKernel() {
+				return new float[] { -1, -1, -1, -1, 8, -1, -1, -1, -1 };
 			}
 
 			@Override
@@ -88,20 +77,16 @@ public class KernelFactory {
 			public int[] getOrigin() {
 				return new int[] { 1, 1 };
 			}
-			
+
 		};
 	}
-	
+
 	public IKernel SHARPEN() {
 		return new IKernel() {
 
 			@Override
-			public double[] getKernel() {
-				return new double[] {
-					 0, -1,  0,
-					-1,  5, -1,
-					 0, -1,  0
-				};
+			public float[] getKernel() {
+				return new float[] { 0, -1, 0, -1, 5, -1, 0, -1, 0 };
 			}
 
 			@Override
@@ -118,20 +103,16 @@ public class KernelFactory {
 			public int[] getOrigin() {
 				return new int[] { 1, 1 };
 			}
-			
+
 		};
 	}
-	
+
 	public IKernel EMBOSS() {
 		return new IKernel() {
 
 			@Override
-			public double[] getKernel() {
-				return new double[] {
-						-2, -1, 0,
-						-1,  1, 1,
-						 0,  1, 2
-				};
+			public float[] getKernel() {
+				return new float[] { -2, -1, 0, -1, 1, 1, 0, 1, 2 };
 			}
 
 			@Override
@@ -148,15 +129,16 @@ public class KernelFactory {
 			public int[] getOrigin() {
 				return new int[] { 1, 1 };
 			}
-			
+
 		};
 	}
-	
-	public IKernel CUSTOM_NORMALIZED_KERNEL(final double[] kernel, final int width, final int height, final int[] origin) {
+
+	public IKernel CUSTOM_NORMALIZED_KERNEL(final float[] kernel, final int width, final int height,
+			final int[] origin) {
 		return new IKernel() {
 
 			@Override
-			public double[] getKernel() {
+			public float[] getKernel() {
 				return kernel;
 			}
 
@@ -174,33 +156,35 @@ public class KernelFactory {
 			public int[] getOrigin() {
 				return origin;
 			}
-			
+
 		};
 	}
-	
-	public IKernel CUSTOM_UNNORMALIZED_KERNEL(final double[] kernel, final int width, final int height, final int[] origin) {
+
+	public IKernel CUSTOM_UNNORMALIZED_KERNEL(final float[] kernel, final int width, final int height,
+			final int[] origin) {
 		return CUSTOM_NORMALIZED_KERNEL(Normalize(kernel), width, height, origin);
 	}
-	
-	public IKernel CUSTOM_GENERATED_KERNEL(final int originValue, final int otherValue, final int width, final int height, final int[] origin) {
-		double[] kernel = new double[width * height];
+
+	public IKernel CUSTOM_GENERATED_KERNEL(final int originValue, final int otherValue, final int width,
+			final int height, final int[] origin) {
+		float[] kernel = new float[width * height];
 		for (int npos = 0; npos < kernel.length; ++npos)
 			kernel[npos] = otherValue;
 		kernel[origin[0] * width + origin[1]] = originValue;
 		return CUSTOM_UNNORMALIZED_KERNEL(kernel, width, height, origin);
 	}
-	
-	private double[] Normalize(final double[] kernel) {
+
+	private float[] Normalize(final float[] kernel) {
 		int sum = 0;
-		
+
 		for (int npos = 0; npos < kernel.length; ++npos) {
 			sum += kernel[npos];
 		}
-		
+
 		for (int npos = 0; npos < kernel.length; ++npos) {
-			kernel[npos] /= (double)sum;
+			kernel[npos] /= (float) sum;
 		}
-		
+
 		return kernel;
 	}
 }
